Update toolchain and refactor `report` method
First issue:
The expanded code of `info!` macro changes when the `log` feature is
enable on the `tracing` crate. The expanded code performs a `move` of
the input parameters.
Said that, you will receive a compilation error if you are passing a
param which calls a `move` by itself (unless the param supports the
`Copy` trait)

For futher information related to the compilation error see:
https://github.com/dusk-network/kadcast/issues/60

Second issue:
I suspect that `Cargo` is behaving wrong, because it enables a nested
dependency feature even if this feature is not explicit declared.
This happens for both `edition = "2021"` and `resolver = "2"`.
Indeed, in the case of the original issue, the `tonic` crate add the
`tower` traversal dependency which use the `tracing` crate with the
`log` feature enabled.

For futher information related to tower dependency see:
https://github.com/tower-rs/tower/blob/b20b3cbd5788727996510fb2e8db04623f60f8ee/tower/Cargo.toml#L27

Refactoring this way, we are sure there will be at most only one `move`

Resolves #60